Dynamic-access coverage for io.smallrye.config.ConfigMappingLoader$ConfigMappingImplementation (13/28)
