+++ tests/src/io.smallrye.config/smallrye-config-core/3.16.0/user-code-filter.json
@@ -1,10 +1,13 @@
 {
-  "rules": [
+  "rules" : [
     {
-      "excludeClasses": "**"
+      "excludeClasses" : "**"
     },
     {
-      "includeClasses": "io.smallrye.config.**"
+      "includeClasses" : "io.smallrye.config.**"
+    },
+    {
+      "includeClasses" : "io_smallrye_config.smallrye_config_core.**"
     }
   ]
 }
